# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  name: string;
  title: string;
  location: string;
  bio: string;
  githubUrl: string;
  linkedinUrl: string;
  email: string;
  photoUrl: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  language?: string | null;
  githubUrl: string;
  stars: number;
  featured: boolean;
  tags: string[];
  updatedAt: string;
}

export interface SkillCategory {
  id: number;
  category: string;
  skills: string[];
}

export interface ContactRequest {
  name: string;
  email: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}