# lib/api-zod/src/generated/types/contactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactRequest {
  name: string;
  email: string;
  message: string;
}